feat: normalize tag storage to song_tags table with multi-valued tag support
Replace individual Option<String> fields on Song struct with a normalized
Vec<(String, String)> tags collection backed by a dedicated song_tags table.
This enables multi-valued tags (e.g., multiple artists, labels, musicbrainz IDs)
matching MPD's native behavior.

- Rewrite Song struct with tag()/tag_values()/tag_with_fallback() API
- Add song_tags table with (song_id, tag, value) schema
- Batch tag loading for efficient database queries
- Update filter SQL to use EXISTS subqueries on song_tags
- Multi-valued tag support in list, count, find commands
- Update all protocol response formatting, test assertions, and examples

+++ rmpd-protocol/src/commands/playback.rs
@@ -53,7 +53,7 @@ pub async fn handle_play_command(state: &AppState, position: Option<u32>) -> Str
                 status.audio_format = Some(rmpd_core::song::AudioFormat {
                     sample_rate: sr,
                     channels: ch,
-                    bits_per_sample: bps,
+                    bits_per_sample: bps as u8,
                 });
             }
